Fixed Corrupted Files and Added room types
Several GameMaker objects' events were overwritten with those of other
objects. Used the git to recover the old events and fixed a few bugs for
those events in the process. Added what should be all of the room types.
They include Rooms with 2 and 4 spaces, and four permutations of rooms
with 3 spaces.